+++ config/responsecache.php
@@ -2,8 +2,7 @@
 
 return [
     /*
-     *  This is the master switch to enable of disable the response cache. If set to
-     *  false no responses will be cached.
+     * Determine if the response cache middleware should be enabled.
      */
     'enabled' => env('RESPONSE_CACHE_ENABLED', true),
 
@@ -14,25 +13,25 @@ return [
      *  You can provide your own class given that it implements the
      *  CacheProfile interface.
      */
-    'cacheProfile' => Spatie\ResponseCache\CacheProfiles\CacheAllSuccessfulGetRequests::class,
+    'cache_profile' => Spatie\ResponseCache\CacheProfiles\CacheAllSuccessfulGetRequests::class,
 
     /*
      * When using the default CacheRequestFilter this setting controls the
      * number of minutes responses must be cached.
      */
-    'cacheLifetimeInMinutes' => env('RESPONSE_CACHE_LIFETIME', 60 * 24 * 7),
+    'cache_lifetime_in_minutes' => env('RESPONSE_CACHE_LIFETIME', 60 * 24 * 7),
 
     /*
      * This setting determines if a http header named "Laravel-responsecache"
      * with the cache time should be added to a cached response. This
      * can be handy when debugging.
      */
-    'addCacheTimeHeader' => env('APP_DEBUG', true),
+    'add_cache_time_header' => env('APP_DEBUG', true),
 
     /*
      * Here you may define the cache store that should be used to store
      * requests. This can be the name of any store that is
      * configured in app/config/cache.php
      */
-    'cacheStore' => env('RESPONSE_CACHE_DRIVER', 'file'),
+    'cache_store' => env('RESPONSE_CACHE_DRIVER', 'file'),
 ];