+++ src/CacheProfiles/CacheAllSuccessfulGetRequests.php
@@ -5,16 +5,9 @@ namespace Spatie\ResponseCache\CacheProfiles;
 use Illuminate\Http\Request;
 use Symfony\Component\HttpFoundation\Response;
 
-class CacheAllSuccessfulGetRequests extends BaseCacheProfile implements CacheProfile
+class CacheAllSuccessfulGetRequests extends BaseCacheProfile
 {
-    /**
-     * Determine if the given request should be cached;.
-     *
-     * @param \Illuminate\Http\Request $request
-     *
-     * @return bool
-     */
-    public function shouldCacheRequest(Request $request)
+    public function shouldCacheRequest(Request $request): bool
     {
         if ($request->ajax()) {
             return false;
@@ -27,14 +20,7 @@ class CacheAllSuccessfulGetRequests extends BaseCacheProfile implements CachePro
         return $request->isMethod('get');
     }
 
-    /**
-     * Determine if the given response should be cached.
-     *
-     * @param \Symfony\Component\HttpFoundation\Response $response
-     *
-     * @return bool
-     */
-    public function shouldCacheResponse(Response $response)
+    public function shouldCacheResponse(Response $response): bool
     {
         return $response->isSuccessful() || $response->isRedirection();
     }