+++ src/CacheProfiles/CacheProfile.php
@@ -2,47 +2,37 @@
 
 namespace Spatie\ResponseCache\CacheProfiles;
 
+use DateTime;
 use Illuminate\Http\Request;
 use Symfony\Component\HttpFoundation\Response;
 
 interface CacheProfile
 {
-    /**
+    /*
+     * Determine if the response cache middleware should be enabled.
+     */
+    public function enabled(Request $request): bool;
+
+    /*
      * Determine if the given request should be cached.
-     *
-     * @param \Illuminate\Http\Request $request
-     *
-     * @return bool
      */
-    public function shouldCacheRequest(Request $request);
+    public function shouldCacheRequest(Request $request): bool;
 
-    /**
+    /*
      * Determine if the given response should be cached.
-     *
-     * @param \Symfony\Component\HttpFoundation\Response $response
-     *
-     * @return bool
      */
-    public function shouldCacheResponse(Response $response);
+    public function shouldCacheResponse(Response $response): bool;
 
-    /**
+    /*
      * Return the time when the cache must be invalidated.
-     *
-     * @param \Illuminate\Http\Request $request
-     *
-     * @return \DateTime
      */
-    public function cacheRequestUntil(Request $request);
+    public function cacheRequestUntil(Request $request): DateTime;
 
-    /**
+    /*
      * Return a string to differentiate this request from others.
      *
      * For example: if you want a different cache per user you could return the id of
      * the logged in user.
-     *
-     * @param \Illuminate\Http\Request $request
-     *
-     * @return string
      */
-    public function cacheNameSuffix(Request $request);
+    public function cacheNameSuffix(Request $request): string;
 }