+++ src/ResponseCache.php
@@ -17,11 +17,6 @@ class ResponseCache
     /** @var CacheProfile */
     protected $cacheProfile;
 
-    /**
-     * @param \Spatie\ResponseCache\ResponseCacheRepository    $cache
-     * @param \Spatie\ResponseCache\RequestHasher              $hasher
-     * @param \Spatie\ResponseCache\CacheProfiles\CacheProfile $cacheProfile
-     */
     public function __construct(ResponseCacheRepository $cache, RequestHasher $hasher, CacheProfile $cacheProfile)
     {
         $this->cache = $cache;
@@ -29,94 +24,61 @@ class ResponseCache
         $this->cacheProfile = $cacheProfile;
     }
 
-    /**
-     * Determine if the given request should be cached.
-     *
-     * @param \Illuminate\Http\Request                   $request
-     * @param \Symfony\Component\HttpFoundation\Response $response
-     *
-     * @return bool
-     */
-    public function shouldCache(Request $request, Response $response)
+    public function enabled(Request $request): bool
     {
-        if (!config('laravel-responsecache.enabled')) {
+        return $this->cacheProfile->enabled($request);
+    }
+
+    public function shouldCache(Request $request, Response $response): bool
+    {
+        if ($request->attributes->has('responsecache.doNotCache')) {
             return false;
         }
 
-        if ($request->attributes->has('laravel-cacheresponse.doNotCache')) {
-            return false;
-        }
-
-        if (!$this->cacheProfile->shouldCacheRequest($request)) {
+        if (! $this->cacheProfile->shouldCacheRequest($request)) {
             return false;
         }
 
         return $this->cacheProfile->shouldCacheResponse($response);
     }
 
-    /**
-     * Store the given response in the cache.
-     *
-     * @param \Illuminate\Http\Request                   $request
-     * @param \Symfony\Component\HttpFoundation\Response $response
-     */
-    public function cacheResponse(Request $request, Response $response)
+    public function cacheResponse(Request $request, Response $response): Response
     {
-        if (config('laravel-responsecache.addCacheTimeHeader')) {
+        if (config('responsecache.add_cache_time_header')) {
             $response = $this->addCachedHeader($response);
         }
 
-        $this->cache->put($this->hasher->getHashFor($request), $response, $this->cacheProfile->cacheRequestUntil($request));
+        $this->cache->put(
+            $this->hasher->getHashFor($request),
+            $response,
+            $this->cacheProfile->cacheRequestUntil($request)
+        );
+
+        return $response;
     }
 
-    /**
-     * Determine if the given request has been cached.
-     *
-     * @param \Illuminate\Http\Request $request
-     *
-     * @return bool
-     */
-    public function hasCached(Request $request)
+    public function hasBeenCached(Request $request): bool
     {
-        if (!config('laravel-responsecache.enabled')) {
-            return false;
-        }
-
-        return $this->cache->has($this->hasher->getHashFor($request));
+        return config('responsecache.enabled')
+            ? $this->cache->has($this->hasher->getHashFor($request))
+            : false;
     }
 
-    /**
-     * Get the cached response for the given request.
-     *
-     * @param \Illuminate\Http\Request $request
-     *
-     * @return \Symfony\Component\HttpFoundation\Response
-     */
-    public function getCachedResponseFor(Request $request)
+    public function getCachedResponseFor(Request $request): Response
     {
         return $this->cache->get($this->hasher->getHashFor($request));
     }
 
-    /**
-     *  Flush the cache.
-     */
     public function flush()
     {
         $this->cache->flush();
     }
 
-    /**
-     * Add a header with the cache date on the response.
-     *
-     * @param \Symfony\Component\HttpFoundation\Response $response
-     *
-     * @return \Symfony\Component\HttpFoundation\Response
-     */
-    protected function addCachedHeader(Response $response)
+    protected function addCachedHeader(Response $response): Response
     {
         $clonedResponse = clone $response;
 
-        $clonedResponse->headers->set('Laravel-reponsecache', 'cached on '.date('Y-m-d H:i:s'));
+        $clonedResponse->headers->set('laravel-responsecache', 'cached on '.date('Y-m-d H:i:s'));
 
         return $clonedResponse;
     }