+++ src/ResponseCacheRepository.php
@@ -2,8 +2,8 @@
 
 namespace Spatie\ResponseCache;
 
-use Illuminate\Contracts\Config\Repository as Repository;
-use Illuminate\Contracts\Foundation\Application;
+use Illuminate\Cache\Repository;
+use Symfony\Component\HttpFoundation\Response;
 
 class ResponseCacheRepository
 {
@@ -13,45 +13,28 @@ class ResponseCacheRepository
     /** @var \Spatie\ResponseCache\ResponseSerializer */
     protected $responseSerializer;
 
-    /** @var string */
-    protected $cacheStoreName;
-
-    /**
-     * @param \Illuminate\Contracts\Foundation\Application $app
-     * @param \Spatie\ResponseCache\ResponseSerializer     $responseSerializer
-     */
-    public function __construct(Application $app, ResponseSerializer $responseSerializer, Repository $config)
+    public function __construct(ResponseSerializer $responseSerializer, Repository $cache)
     {
-        $this->cache = $app['cache']->store($config->get('laravel-responsecache.cacheStore'));
+        $this->cache = $cache;
         $this->responseSerializer = $responseSerializer;
     }
 
     /**
-     * @param string                                     $key
+     * @param string $key
      * @param \Symfony\Component\HttpFoundation\Response $response
-     * @param \DateTime|int                              $minutes
+     * @param \DateTime|int $minutes
      */
-    public function put($key, $response, $minutes)
+    public function put(string $key, $response, $minutes)
     {
         $this->cache->put($key, $this->responseSerializer->serialize($response), $minutes);
     }
 
-    /**
-     * @param string $key
-     *
-     * @return bool
-     */
-    public function has($key)
+    public function has(string $key): bool
     {
         return $this->cache->has($key);
     }
 
-    /**
-     * @param string $key
-     *
-     * @return \Symfony\Component\HttpFoundation\Response
-     */
-    public function get($key)
+    public function get(string $key): Response
     {
         return $this->responseSerializer->unserialize($this->cache->get($key));
     }