+++ tests/Commands/FlushCommandTest.php
@@ -2,17 +2,20 @@
 
 namespace Spatie\ResponseCache\Test\Commands;
 
+use Event;
 use Illuminate\Support\Facades\Artisan;
 use Spatie\ResponseCache\Test\TestCase;
+use Spatie\ResponseCache\Events\FlushedResponseCache;
+use Spatie\ResponseCache\Events\FlushingResponseCache;
 
-class ClearCommandTest extends TestCase
+class FlushCommandTest extends TestCase
 {
     /** @test */
     public function it_will_clear_the_cache()
     {
         $firstResponse = $this->call('GET', '/random');
 
-        Artisan::call('responsecache:clear');
+        Artisan::call('responsecache:flush');
 
         $secondResponse = $this->call('GET', '/random');
 
@@ -20,5 +23,16 @@ class ClearCommandTest extends TestCase
         $this->assertRegularResponse($secondResponse);
 
         $this->assertDifferentResponse($firstResponse, $secondResponse);
+    }
+
+    /** @test */
+    public function it_will_fire_events_when_clearing_the_cache()
+    {
+        Event::fake();
+
+        Artisan::call('responsecache:flush');
+
+        Event::assertDispatched(FlushingResponseCache::class);
+        Event::assertDispatched(FlushedResponseCache::class);
     }
 }